Stage 2: long -> int32_t / unsigned long -> uint32_t sweep
Mechanical width-explicit rename across the public API, its wrappers,
the engine, kplib, and the oracle. On 32-bit MSVC (LLP64) `long` and
`int32_t` are both 4 bytes so this is hash-neutral; the goldens must
not move. On 64-bit LP64 targets (future Linux/macOS) `long` would be
8 bytes, so fixing the type name now is a prerequisite for anything
downstream — struct layouts, binary file formats, and the ABI to the
upcoming Rust bindings all need known-width fields.

Order matters: `signed long` -> int32_t and `unsigned long` -> uint32_t
were applied before the bare `long` -> int32_t pass so those
compounds wouldn't end up as `signed int32_t` / `unsigned int32_t`.

Out of scope (Step 5): pointer-cast sites of the form (long)ptr,
which should become (intptr_t)ptr. They are width-identical to
(int32_t)ptr on 32-bit MSVC so they ride along here harmlessly; the
next commit flips them to intptr_t so they're correct on 64-bit too.

Preserved as-is: (unsigned long long)h casts in the oracle's %llx
fprintfs — that's the C-standard-idiomatic type for %llx and does
not want to become (uint64_t), which may typedef differently on MSVC.
Identifiers containing "long" as a substring (pow2long, longpal,
templongbuf) are preserved by word-bounded regex.

+++ tests/oracle/oracle.c
@@ -18,7 +18,7 @@
  * global, which is not exported from voxlap.dll. That keeps this harness
  * free of any engine modifications.
  *
- * Note: voxsetframebuffer's first arg is typed `long`, which on Win32 x86
+ * Note: voxsetframebuffer's first arg is typed `int32_t`, which on Win32 x86
  * is the same width as a pointer but NOT on x64. Stage 2 widens this API.
  */
 
@@ -32,13 +32,13 @@
 
 /* Wrappers declared in exports.c (not in voxlap5.h). Re-declare here so the
  * oracle doesn't need a second header. Linkage matches the VOXLAP_API defs. */
-extern void setRectOneColor(lpoint3d *hit1, lpoint3d *hit2, long ARGB);
-extern void set_curcol(long v);
-extern void set_jitamount(long v);
-extern void set_colfunc(long (*v)(lpoint3d *));
-extern void set_fogcol(long v);
-extern void set_anginc(long v);
-extern void setMaxScanDist(long v);
+extern void setRectOneColor(lpoint3d *hit1, lpoint3d *hit2, int32_t ARGB);
+extern void set_curcol(int32_t v);
+extern void set_jitamount(int32_t v);
+extern void set_colfunc(int32_t (*v)(lpoint3d *));
+extern void set_fogcol(int32_t v);
+extern void set_anginc(int32_t v);
+extern void setMaxScanDist(int32_t v);
 
 /* Voxlap's 32-bit colour is packed as (brightness<<24) | (R<<16) | (G<<8) | B.
  * The alpha byte is *brightness*, not opacity: setting it to 0x00 renders
@@ -72,7 +72,7 @@ static void build_scene(void) {
 	 * large enough to cover the diagonal of our 448-voxel playable
 	 * box (~633) with headroom. */
 	setMaxScanDist(1024);
-	set_fogcol((long)BR(0x87ceeb));
+	set_fogcol((int32_t)BR(0x87ceeb));
 
 	setsideshades(0, 0, 0, 0, 0, 0);
 	set_colfunc(curcolfunc);
@@ -81,14 +81,14 @@ static void build_scene(void) {
 
 	/* Carve a 448x448x185 playable box at map centre; the carve paints
 	 * exposed inside faces of the surrounding solid with curcol. */
-	set_curcol((long)BR(0x87ceeb));
+	set_curcol((int32_t)BR(0x87ceeb));
 	a.x = 800;   a.y = 800;   a.z = 5;
 	b.x = 1248;  b.y = 1248;  b.z = 189;
 	setrect(&a, &b, -1);
 
 	/* Grey floor slab inside the playable box so looking down reads
 	 * as stone. Top surface at z=185 is what the camera sees. */
-	set_curcol((long)BR(0x606878));
+	set_curcol((int32_t)BR(0x606878));
 	a.x = 800;   a.y = 800;   a.z = 185;
 	b.x = 1248;  b.y = 1248;  b.z = 189;
 	setrect(&a, &b, 0);
@@ -96,20 +96,20 @@ static void build_scene(void) {
 	/* Red pillar */
 	a.x = 1010; a.y = 1090; a.z = 155;
 	b.x = 1020; b.y = 1100; b.z = 184;
-	setRectOneColor(&a, &b, (long)BR(0xff3030));
+	setRectOneColor(&a, &b, (int32_t)BR(0xff3030));
 
 	/* Green cube */
 	a.x = 1030; a.y = 1050; a.z = 175;
 	b.x = 1040; b.y = 1060; b.z = 184;
-	setRectOneColor(&a, &b, (long)BR(0x30c030));
+	setRectOneColor(&a, &b, (int32_t)BR(0x30c030));
 
 	/* Blue flat tile */
 	a.x = 1000; a.y = 1030; a.z = 183;
 	b.x = 1050; b.y = 1070; b.z = 184;
-	setRectOneColor(&a, &b, (long)BR(0x3060ff));
+	setRectOneColor(&a, &b, (int32_t)BR(0x3060ff));
 
 	/* Yellow sphere */
-	set_curcol((long)BR(0xffd050));
+	set_curcol((int32_t)BR(0xffd050));
 	c.x = 1060; c.y = 1040; c.z = 178;
 	setsphere(&c, 8, 0);
 
@@ -153,7 +153,7 @@ int main(void) {
 	if (initvoxlap() < 0) { fprintf(stderr, "initvoxlap failed\n"); return 1; }
 
 	build_scene();
-	voxsetframebuffer((long)(intptr_t)g_fb, BYTESPERLINE, XRES, YRES);
+	voxsetframebuffer((int32_t)(intptr_t)g_fb, BYTESPERLINE, XRES, YRES);
 
 	hf = fopen("hashes.txt", "w");
 	if (!hf) { fprintf(stderr, "cannot open hashes.txt\n"); return 2; }
